Update MultiDropdown to Dropdown
