+++ src/components/Dropdown/index.ts
@@ -1,0 +1,1 @@
+export { default as Dropdown } from './Dropdown';
