--- src/components/MultiDropdown/index.ts
@@ -1,1 +1,0 @@
-export * from './MultiDropdown';
